Activity.cs model class handles weight now. ActivitiesController.cs, ActivitiesApiCall.cs, ActivityRepository.cs, IActivityRepository.cs and JsonProcessor.cs modified too

+++ HealthManagerServer/HealthManagerServer/Service/ExternalApis/ActivitiesApiCall.cs
@@ -9,20 +9,16 @@ public class ActivitiesApiCall
         _apiKey = "pe8bDjzt2qs1AeJNbzukUw==sXVqA6FU8t4IBEFG";
     }
 
-    public async Task<string> GetActivitiesData(string activityName, double weight = 160, int duration=60 )
+    public async Task<string> GetActivitiesData(string activityName, int weight = 160, int duration=60 )
     {
         string url = $"https://api.api-ninjas.com/v1/caloriesburned/?activity={activityName}&weight={weight}&duration={duration}";
         using var client = new HttpClient();
         client.DefaultRequestHeaders.Add("X-Api-Key", _apiKey);
         var response = await client.GetAsync(url);
         
-        
-
         if (response.IsSuccessStatusCode)
         {
             var content = await response.Content.ReadAsStringAsync();
-            
-
             return content;
         }
